+++ CHANGELOG.md
@@ -95,6 +95,7 @@
 * [#4354](https://github.com/bbatsov/rubocop/pull/4354): Add autocorrect to `Style/FormatString`. ([@hoshinotsuyoshi][])
 * [#4021](https://github.com/bbatsov/rubocop/pull/4021): Add new `Style/MultipleComparison` cop. ([@dabroz][])
 * New `Lint/RescueType` cop. ([@rrosenblum][])
+* [#4355](https://github.com/bbatsov/rubocop/pull/4355): Add `rubocop:todo` directive alongside `rubocop:disable`. ([@vergenzt][])
 
 ### Changes